(feat) Use NACLs to lock down load balancers
It was misunderstood how NLBs work when instances are added by IP (as opposed
to instance ID):

  - The NLBs cannot have a security group associated with them
  - Security groups on the application only block/allow communication based on
    the IP of the NLB, rather than client IP.

This doesn't leave any way to have an IP whitelist for the application, unless
we use network access control lists. Conveniently, each NLB is in its own
subnet, with nothing else in it, so NACLs on those subnets only affect the
NLBs.